using decoder in segw

+++ nodes/mqtt/src/controller.cpp
@@ -39,7 +39,7 @@ namespace node
 	cyng::vector_t controller::create_config(std::fstream& fout, boost::filesystem::path&& tmp, boost::filesystem::path&& cwd) const
 	{
 		boost::random::mt19937 rng_;
-        rng_.seed(std::time(nullptr));
+        rng_.seed(static_cast<std::uint32_t>(std::time(nullptr)));
 		boost::random::uniform_int_distribution<int> monitor_dist(10, 120);
 
 		return cyng::vector_factory({
